refactor:setting typeorm improved
