+++ src/modules/workspace_keys/entity/workspace_keys.entity.ts
@@ -2,9 +2,9 @@ import { Entity, Column, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
 export class WorkSpaceKeys {
-    @PrimaryGeneratedColumn()
-    id: number;
+  @PrimaryGeneratedColumn()
+  id: number;
 
-    @Column()
-    value: string;
+  @Column()
+  value: string;
 }
